[Refactor] 채팅방 참여자 추가 기능
