+++ frontend/src/services/websocketService.js
@@ -1,5 +1,5 @@
-import { useChatRoomStore } from '@/stores/socket/chat/useChatRoomStore';
-import { useChatMessageStore } from "@/stores/socket/chat/useChatMessageStore";
+import { useChatRoomStore } from '@/stores/chat/useChatRoomStore';
+import { useChatMessageStore } from "@/stores/chat/useChatMessageStore";
 import { useNotificationStore } from '@/stores/socket/useAlarmStore';
 import { useDocsEditStore } from '@/stores/socket/useDocsEditStore';
 import { Client } from '@stomp/stompjs';
